+++ app/ide-desktop/lib/client/src/project-management.ts
@@ -12,21 +12,23 @@ import fss from 'node:fs'
 import electron from 'electron'
 import crypto from 'node:crypto'
 
-import { isProjectRoot, projectMetadataRelative } from './paths'
+import { isProjectRoot, projectMetadataRelative } from './paths.js'
 import fsSync from 'node:fs'
-import { BUNDLED_PROJECT_EXTENSION } from '../paths'
 import * as tar from 'tar'
+import { getCommonPrefix } from '../../../utils.js'
+import { logger } from 'enso-content-config'
+import {BUNDLED_PROJECT_EXTENSION} from "../file-associations";
 
-interface ProjectMetadata {
-    id: string
-}
+// =======================
+// === Project Import ===
+// =======================
 
 /** Open a project from the given path. Path can be either a source file under the project root, or the project
  * bundle. If needed, the project will be imported into the Project Manager-enabled location.
  *
  * @returns Project ID (from Project Manager's metadata) identifying the imported project.
  */
-export function openFile(openedPath: string): string {
+export function importProjectFromPath(openedPath: string): string {
     if (path.extname(openedPath).endsWith(BUNDLED_PROJECT_EXTENSION)) {
         // The second part of condition is for the case when someone names a directory like `my-project.enso-project`
         // and stores the project there. Not the most fortunate move, but...
@@ -37,46 +39,15 @@ export function openFile(openedPath: string): string {
             return importBundle(openedPath)
         }
     } else {
-        console.log(`Opening file: ${openedPath}`)
+        logger.log(`Opening file: '${openedPath}'.`)
         const rootPath = getProjectRoot(openedPath)
         // Check if the project root is under the projects directory. If it is, we can open it.
         // Otherwise, we need to install it first.
         if (rootPath === null) {
-            // Perhaps bundle?
             let message = `File '${openedPath}' does not belong to the Enso project.`
             throw new Error(message)
         }
         return importDirectory(rootPath)
-    }
-}
-
-/** Get the common prefix of the two strings. */
-export function getCommonPrefix(a: string, b: string): string {
-    let i = 0
-    while (i < a.length && i < b.length && a[i] === b[i]) {
-        i++
-    }
-    return a.slice(0, i)
-}
-
-/** If the bundle consists of a single directory, return its name. Otherwise, return null. */
-export function bundleRootDirectory(bundlePath: string): string | null {
-    // We need to look up the root directory among the tarball entries.
-    let commonPrefix: string | null = null
-    tar.list({
-        file: bundlePath,
-        sync: true,
-        onentry: entry => {
-            // We normalize to get rid of leading `.` (if any).
-            let path = entry.path.normalize()
-            commonPrefix = commonPrefix === null ? path : getCommonPrefix(commonPrefix, path)
-            console.log(`Entry: ${entry.path}, common prefix: ${commonPrefix}`)
-        },
-    })
-    if (commonPrefix !== null && commonPrefix !== '') {
-        return path.basename(commonPrefix)
-    } else {
-        return null
     }
 }
 
@@ -87,9 +58,7 @@ export function bundleRootDirectory(bundlePath: string): string | null {
 export function importBundle(bundlePath: string): string {
     // The bundle is a tarball, so we just need to extract it to the right location.
     const bundleRoot = bundleRootDirectory(bundlePath)
-    const targetDirectory = bundleRoot
-        ? generateDirectoryName(bundleRoot)
-        : generateDirectoryName(bundlePath)
+    const targetDirectory = generateDirectoryName(bundleRoot ?? bundlePath)
     fss.mkdirSync(targetDirectory, { recursive: true })
     // To be more resilient against different ways that user might attempt to create a bundle, we try to support
     // both archives that:
@@ -116,30 +85,100 @@ export function importBundle(bundlePath: string): string {
 export function importDirectory(rootPath: string): string {
     if (isProjectInstalled(rootPath)) {
         // Project is already visible to Project Manager, so we can just return its ID.
-        console.debug(`Project already installed: ${rootPath}`)
+        logger.log(`Project already installed: '${rootPath}'.`)
         return getProjectId(rootPath)
     }
 
-    console.debug(`Importing a project copy from: ${rootPath}`)
+    logger.log(`Importing a project copy from: '${rootPath}'.`)
     const targetDirectory = generateDirectoryName(rootPath)
     if (fsSync.existsSync(targetDirectory)) {
         const message = `Project directory already exists: ${targetDirectory}.`
         throw new Error(message)
     }
 
-    console.debug(`Copying: ${rootPath} -> ${targetDirectory}`)
+    logger.log(`Copying: '${rootPath}' -> '${targetDirectory}'.`)
     fsSync.cpSync(rootPath, targetDirectory, { recursive: true })
     // Update the project ID, so we are certain that it is unique. This would be violated, if we imported the same
     // project multiple times.
     return updateId(targetDirectory)
 }
 
+
+// ================
+// === Metadata ===
+// ================
+
+/** The Project Manager's metadata associated with a project.
+ *
+ * The property list is not exhaustive, it only contains the properties that we need.
+ **/
+interface ProjectMetadata {
+    /** The ID of the project. It is only used in communication with project manager, it has no semantic meaning. */
+    id: string
+}
+
+/** Get the ID from the project metadata. */
+export function getProjectId(projectRoot: string): string {
+    return getMetadata(projectRoot).id
+}
+
+/** Retrieve the project's metadata. */
+export function getMetadata(projectRoot: string): ProjectMetadata {
+    const metadataPath = path.join(projectRoot, projectMetadataRelative)
+    const jsonText = fss.readFileSync(metadataPath, 'utf8')
+    return JSON.parse(jsonText)
+}
+
+/** Write the project's metadata. */
+export function writeMetadata(projectRoot: string, metadata: ProjectMetadata): void {
+    const metadataPath = path.join(projectRoot, projectMetadataRelative)
+    fss.writeFileSync(metadataPath, JSON.stringify(metadata, null, 4))
+}
+
+/** Update project's metadata. If the provided updater does not return anything, the metadata file is left intact.
+ *
+ *  The updater function-returned metadata is passed over.
+ **/
+export function updateMetadata(
+    projectRoot: string,
+    updater: (initialMetadata: ProjectMetadata) => ProjectMetadata
+): ProjectMetadata {
+    const metadata = getMetadata(projectRoot)
+    const updatedMetadata = updater(metadata)
+    writeMetadata(projectRoot, updatedMetadata)
+    return updatedMetadata
+}
+
+// =========================
+// === Project Directory ===
+// =========================
+
+/** Check if this bundle is a compressed directory (rather than directly containing the project files). If it is, we
+ *  return the name of the directory. Otherwise, we return `null`. */
+export function bundleRootDirectory(bundlePath: string): string | null {
+    // We need to look up the root directory among the tarball entries.
+    let commonPrefix: string | null = null
+    tar.list({
+        file: bundlePath,
+        sync: true,
+        onentry: entry => {
+            // We normalize to get rid of leading `.` (if any).
+            let path = entry.path.normalize()
+            commonPrefix = commonPrefix === null ? path : getCommonPrefix(commonPrefix, path)
+        },
+    })
+    if (commonPrefix !== null && commonPrefix !== '') {
+        return path.basename(commonPrefix)
+    } else {
+        return null
+    }
+}
+
 /** Generate a name for project using given base string. Suffixes are added if there's a collision.
  *
  * For example 'Name' will become 'Name_1' if there's already a directory named 'Name'.
  * If given a name like 'Name_1' it will become 'Name_2' if there's already a directory named 'Name_1'.
- * If a path containing multiple components is given, only the last component is used for the name.
- **/
+ * If a path containing multiple components is given, only the last component is used for the name. */
 export function generateDirectoryName(name: string): string {
     // Use only the last path component.
     name = path.parse(name).name
@@ -158,47 +197,15 @@ export function generateDirectoryName(name: string): string {
             projectsDirectory,
             `${name}${suffix === 0 ? '' : '_' + suffix}`
         )
-        console.log(`Checking if ${candidatePath} exists?`)
         if (!fss.existsSync(candidatePath)) {
             return candidatePath
         }
     }
     // Unreachable.
 }
-
-/** Get the ID from the project metadata. */
-export function getProjectId(projectRoot: string): string {
-    return getMetadata(projectRoot).id
-}
-
-export function getMetadata(projectRoot: string): ProjectMetadata {
-    const metadataPath = path.join(projectRoot, projectMetadataRelative)
-    const jsonText = fss.readFileSync(metadataPath, 'utf8')
-    return JSON.parse(jsonText)
-}
-
-export function writeMetadata(projectRoot: string, metadata: ProjectMetadata): void {
-    updateMetadata(projectRoot, () => metadata)
-}
-
-/** Update project's metadata. If the provided updater does not return anything, the metadata file is left intact.
- *
- *  The updater function-returned metadata is passed over.
- **/
-export function updateMetadata<R>(projectRoot: string, updater: (a: ProjectMetadata) => R): R {
-    const metadataPath = path.join(projectRoot, projectMetadataRelative)
-    const jsonText = fss.readFileSync(metadataPath, 'utf8')
-    const parsedMetadata = JSON.parse(jsonText)
-    const updatedMetadata = updater(parsedMetadata)
-    if (updatedMetadata !== undefined) {
-        fss.writeFileSync(metadataPath, JSON.stringify(updatedMetadata, null, 4))
-    }
-    return updatedMetadata
-}
-
-/** Get the project root path from its subtree. */
+/** Takes a path to a file, presumably located in a project's subtree. Returns the path to the project's root directory
+ * or `null` if the file is not located in a project. */
 export function getProjectRoot(subtreePath: string): string | null {
-    console.debug(`Getting project root for ${subtreePath}.`)
     let currentPath = subtreePath
     while (!isProjectRoot(currentPath)) {
         const parent = path.dirname(currentPath)
@@ -223,6 +230,10 @@ export function isProjectInstalled(projectRoot: string): boolean {
     // Should resolve symlinks and relative paths. Normalize before comparison.
     return path.resolve(projectRootParent) === path.resolve(projectsDirectory)
 }
+
+// ==================
+// === Project ID ===
+// ==================
 
 /** Generates a unique UUID for a project. */
 export function generateId(): string {