+++ app/ide-desktop/utils.ts
@@ -45,3 +45,12 @@ export function requireEnvPathExist(name: string) {
         throw Error(`File with path ${value} read from environment variable ${name} is missing.`)
     }
 }
+
+/** Get the common prefix of the two strings. */
+export function getCommonPrefix(a: string, b: string): string {
+    let i = 0
+    while (i < a.length && i < b.length && a[i] === b[i]) {
+        i++
+    }
+    return a.slice(0, i)
+}